Added list builder for geo sphere and new view for geo sphere page
